requirements.txt
alembic==1.12.0
black==23.9.1
fastapi==0.103.1
httpx==0.25.0
passlib[bcrypt]==1.7.4
pillow==10.0.1
pydantic==2.6.3
pylint==2.17.5
pytest==7.4.2
pytest-order==1.1.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
SQLAlchemy==2.0.21
uvicorn==0.23.2
